codegen: extrat FuncDefs and related stuff into function_deffinitions.go
